library/CWS/requirements.txt
pytorch-crf
scikit-learn
transformers==4.0.0